Added the Author table + file formatting changes.
The project will now support an author table to store data of the authors. Each book in the book table will be linked to a author in the data table. This will enable me to create advanced SQL statements and develop an advanced library manager API.

Right now, the get methods are the only functions available. However, I plan to add functionality to the other methods. I've also updated the file structure to have a better file organization. Finally, I've removed the update columns SQL file in favor of a larger starting SQL file.